moved AutoComplete to a Feature
